Remove dead current_run state and consolidate test fixtures
- Remove get_current_run() and state.current_run — no route consumed them
- Add concurrency comment on current_eval (single-eval-at-a-time constraint)
- Extract shared app/client fixtures into tests/conftest.py
