+++ src/midojo/app/dependencies.py
@@ -12,12 +12,6 @@ def get_suite() -> YAMLTaskSuite:
     return state.suite
 
 
-def get_current_run() -> Run:
-    if state.current_run is None:
-        raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="No run in progress.")
-    return state.current_run
-
-
 def get_current_eval() -> Evaluation:
     if state.current_eval is None:
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST, detail="No evaluation in progress.")
